requirements.txt
ccxt
pytz
pandas
datetime
loguru
pyyaml
telegram
schedule
requests